refactor(slack): split gateway socket and dispatch modules
